fix(ci): fix release/prune which runs before make install
